Distrubance Observer V1.1 (With init)

+++ Inc/tmc6200/TMC6200.h
@@ -5,7 +5,7 @@
 #include "TMC6200_Register.h"
 #include "TMC6200_Constants.h"
 
-int tmc6200_readInt(uint8_t drv, uint8_t address);
+int16_t tmc6200_readInt(uint8_t drv, uint8_t address);
 void tmc6200_writeInt(uint8_t drv, uint8_t address, int value);
 
 #endif /* TMC_IC_TMC6630_H_ */
